+++ src/constants/code-editor.ts
@@ -175,6 +175,11 @@ const BASE_CONTEXT_SUGGESTIONS: CodeEditorSuggestionItem[] = [
     detail: 'Grafana properties and methods.',
   },
   {
+    label: 'context.grafana.backendService',
+    kind: CodeEditorSuggestionItemKind.Property,
+    detail: 'Backend service.',
+  },
+  {
     label: 'context.grafana.locationService',
     kind: CodeEditorSuggestionItemKind.Property,
     detail: 'Location service.',
@@ -224,6 +229,11 @@ const BASE_CONTEXT_SUGGESTIONS: CodeEditorSuggestionItem[] = [
     kind: CodeEditorSuggestionItemKind.Method,
     detail: 'Parse the results from /api/ds/query.',
   },
+  {
+    label: 'context.utils.fileToBase64',
+    kind: CodeEditorSuggestionItemKind.Method,
+    detail: 'Converts File to Base64 string.',
+  },
 ];
 
 /**
@@ -249,6 +259,12 @@ export const CODE_EDITOR_SUGGESTIONS: Record<CodeEditorType, CodeEditorSuggestio
       kind: CodeEditorSuggestionItemKind.Property,
       detail: 'Parsed values from the Initial Request.',
     },
+    /**
+     * Context
+     */
+    ...BASE_CONTEXT_SUGGESTIONS.filter((item) =>
+      ['context', 'context.panel.elements', 'context.panel.initial', 'context.utils.fileToBase64'].includes(item.label)
+    ),
   ],
 
   /**
@@ -312,6 +328,16 @@ export const CODE_EDITOR_SUGGESTIONS: Record<CodeEditorType, CodeEditorSuggestio
       kind: CodeEditorSuggestionItemKind.Method,
       detail: 'Interpolate variables.',
     },
+
+    /**
+     * Context
+     */
+    ...BASE_CONTEXT_SUGGESTIONS.filter((item) => ['context', 'context.panel.elements'].includes(item.label)),
+    {
+      label: 'context.grafana.replaceVariables',
+      kind: CodeEditorSuggestionItemKind.Method,
+      detail: 'Interpolate variables.',
+    },
   ],
 
   /**
@@ -325,6 +351,16 @@ export const CODE_EDITOR_SUGGESTIONS: Record<CodeEditorType, CodeEditorSuggestio
     },
     {
       label: 'replaceVariables',
+      kind: CodeEditorSuggestionItemKind.Method,
+      detail: 'Interpolate variables.',
+    },
+
+    /**
+     * Context
+     */
+    ...BASE_CONTEXT_SUGGESTIONS.filter((item) => ['context', 'context.panel.elements'].includes(item.label)),
+    {
+      label: 'context.grafana.replaceVariables',
       kind: CodeEditorSuggestionItemKind.Method,
       detail: 'Interpolate variables.',
     },
@@ -346,6 +382,14 @@ export const CODE_EDITOR_SUGGESTIONS: Record<CodeEditorType, CodeEditorSuggestio
     },
     {
       label: 'replaceVariables',
+      kind: CodeEditorSuggestionItemKind.Method,
+      detail: 'Interpolate variables.',
+    },
+    ...BASE_CONTEXT_SUGGESTIONS.filter((item) =>
+      ['context', 'context.panel.data', 'context.panel.elements'].includes(item.label)
+    ),
+    {
+      label: 'context.grafana.replaceVariables',
       kind: CodeEditorSuggestionItemKind.Method,
       detail: 'Interpolate variables.',
     },